+++ FixtureController.php
@@ -293,8 +293,9 @@ class FixtureController extends \yii\console\controllers\FixtureController
     /**
      * Notifies user that given fixtures template files were not found.
      * @param array $templatesNames
+     * @since 2.0.4
      */
-    private function notifyNotFoundTemplates($templatesNames)
+    protected function notifyNotFoundTemplates($templatesNames)
     {
         $this->stdout("The following fixtures templates were NOT found:\n\n", Console::FG_RED);
 
@@ -307,8 +308,9 @@ class FixtureController extends \yii\console\controllers\FixtureController
 
     /**
      * Notifies user that there was not found any files matching given input conditions.
+     * @since 2.0.4
      */
-    private function notifyNoTemplatesFound()
+    protected function notifyNoTemplatesFound()
     {
         $this->stdout("No fixtures template files matching input conditions were found under the path:\n\n", Console::FG_RED);
         $this->stdout("\t " . Yii::getAlias($this->templatePath) . " \n\n", Console::FG_GREEN);
@@ -317,8 +319,9 @@ class FixtureController extends \yii\console\controllers\FixtureController
     /**
      * Notifies user that given fixtures template files were generated.
      * @param array $templatesNames
+     * @since 2.0.4
      */
-    private function notifyTemplatesGenerated($templatesNames)
+    protected function notifyTemplatesGenerated($templatesNames)
     {
         $this->stdout("The following fixtures template files were generated:\n\n", Console::FG_YELLOW);
 
@@ -329,7 +332,12 @@ class FixtureController extends \yii\console\controllers\FixtureController
         $this->stdout("\n");
     }
 
-    private function notifyTemplatesCanBeGenerated($templatesNames)
+    /**
+     * Notifies user about templates which could be generated.
+     * @param array $templatesNames
+     * @since 2.0.4
+     */
+    protected function notifyTemplatesCanBeGenerated($templatesNames)
     {
         $this->stdout("Template files path: ", Console::FG_YELLOW);
         $this->stdout(Yii::getAlias($this->templatePath) . "\n\n", Console::FG_GREEN);
@@ -346,8 +354,9 @@ class FixtureController extends \yii\console\controllers\FixtureController
      * by the given parameter.
      * @param array $templatesNames template file names to search. If empty then all files will be searched.
      * @return array
+     * @since 2.0.4
      */
-    private function findTemplatesFiles(array $templatesNames = [])
+    protected function findTemplatesFiles(array $templatesNames = [])
     {
         $findAll = ($templatesNames == []);